requirements.txt
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

